Revert "Bug 2047339. r=geckoview-reviewers,nika,ipc-reviewers" for causing Android gtest failures.
This reverts commit c1d8ad3f29e20b1f9f4a6654ee25b82c56acfeed.

+++ widget/android/GeckoProcessManager.cpp
@@ -18,7 +18,7 @@ namespace mozilla {
 
 // static
 void GeckoProcessManager::GetEditableParent(jni::Object::Param aEditableChild,
-                                            int32_t aContentId,
+                                            int64_t aContentId,
                                             int64_t aTabId) {
   nsCOMPtr<nsIWidget> widget = GetWidget(aContentId, aTabId);
   if (RefPtr<nsWindow> window = nsWindow::From(widget)) {
